custom-editor/requirements.txt
spacy
fastapi
uvicorn
openai
pandas
python-dotenv
joblib
